[HOY-28] fix: 찜버튼 Header로 이동(이관현)
