+++ src/shared/components/card/productCard/productButtons.tsx
@@ -9,15 +9,15 @@ interface ProductButtonsProps {
 
 const ProductButtons = ({ isEditable }: ProductButtonsProps) => {
   return (
-    <div className='flex flex-col gap-[15px]'>
-      <Button variant='primary' className='w-full max-w-none rounded-[8px]'>
+    <div className='flex flex-col gap-[15px] md:flex-row'>
+      <Button variant='primary' className='w-full max-w-none rounded-[8px] md:flex-1'>
         리뷰 작성하기
       </Button>
-      <Button variant='secondary' className='w-full max-w-none rounded-[8px]'>
+      <Button variant='secondary' className='w-full max-w-none rounded-[8px] md:flex-1'>
         비교하기
       </Button>
       {isEditable && (
-        <Button variant='tertiary' className='w-full max-w-none rounded-[8px]'>
+        <Button variant='tertiary' className='w-full max-w-none rounded-[8px] md:flex-1'>
           편집하기
         </Button>
       )}
